remove rest of references to icons
