+++ package.json
@@ -24,7 +24,6 @@
   "devDependencies": {
     "@astrojs/image": "0.7.1",
     "@astrojs/react": "1.1.4",
-    "@fortawesome/fontawesome-svg-core": "6.1.1",
     "@fullhuman/postcss-purgecss": "4.1.3",
     "@popperjs/core": "2.11.6",
     "@types/react": "18.0.17",
@@ -43,7 +42,6 @@
     "postcss": "8.4.16",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-icons": "4.3.1",
     "react-markdown": "8.0.2",
     "react-vertical-timeline-component": "3.5.2",
     "sass": "1.54.5",
