+++ src/components/sections/interests/interests.scss
@@ -1,19 +1,3 @@
-@import "../../../styles/bootstrap_overrides";
-
-.vertical-timeline-element-content {
-  border-top: 3px solid $primary;
-}
-
-@media only screen and (max-width: 1169px) {
-  /* .vertical-timeline--animate */
-  .vertical-timeline-element-content.bounce-in {
-    visibility: visible;
-    -webkit-animation: cd-bounce-2-inverse 0.6s;
-    -moz-animation: cd-bounce-2-inverse 0.6s;
-    animation: cd-bounce-2-inverse 0.6s;
-  }
-}
-
 .book-icon {
   color: #0052cc;
 }
